requirements.txt
autogen[retrieve-chat]
python-dotenv
chromadb
sentence-transformers
